requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
resend==1.0.2
python-dateutil==2.8.2
blinker==1.7.0